Proposal to fix #22: improve navigation & more
This is my proposal to fix #22 issue.

I add a TOC for the git cheatsheet, improve the listing of the
commands/recipes and the navigation, with even a link back to the top
(where the TOC is) in every heading.

I think this will be pretty helpful. What do you think about it,
@edgarshurtado?