Add a pair of new git commit commands
- Adds commands for view a diff of changes at the same time we
  write the extended commit message in our editor.
- Also add a command for globally config git for made verbose
  commits (first of previous commands) the default when we use
  the `git commit` command.

This is pretty useful. Till I know this, I used to do this
separately, first a `git diff` and later the commit, or using
two terminal windows for this. This way, you get the same info
in one window, with the diff just below the list of modified files.
This is a killer option :D.